code/requirements.txt
RPi.GPIO
opencv-python-headless
numpy
tensorflow==2.15.0
picamera2